Solution
Binary search over the largest possible answer. Then check if it is possible using dfs. Recursively check if it is possible to give a certain metal y units by first using up all the units the metal already has, and then by checking all of it's ingredients in the recipe to see if they all have at least that remainder left. This is because in order to craft 1 unit of a certain metal, you need at least 1 unit each of all its ingredients.